Change implementation of `getMethodMap` to only return methods defined in the interface itself.
Summary:
Thrift service can be composed of multiple thrift service definitions in a hierarchy, where the final set of methods is the set of all methods from all services in the hierarchy. However there is no polymorphism or method overriding so there should never be a case where a superclass or subclass have a method of the same name.
In current implementation `getMethodMap` returns all the methods of class itself, and of superclasses. After D75417393 was implemented to gather all superclass interfaces that's not necessary. D75417393 adds ugly handling for method name conflicts - we can remove it here, as methods from same class hierarchy wouldn't be returned in `getMethodMap` - hence any duplicate method will be an exception.

Reviewed By: yfeldblum

Differential Revision: D77590706

fbshipit-source-id: 7151e99ff0c0695011c84f1ed9898230756bf3e9