Explicitly set __DynamicallyCallable notice level for package tests
Summary: Using `HH\dynamic_fun` to load a `<<__DynamicallyCallable>>` function from a string causes us to lose soft-package and module warnings; I believe this is just an oversight in the implementation and should be an easy fix. See {D78069737} for what this looks like in practice. For the moment, just set the flag explicitly on these tests so the rest can move on.

Reviewed By: viratyosin

Differential Revision: D78068658

fbshipit-source-id: c394e76f20c0719e82ce84d9c6a816baa9f952ce