Fix Hack LSP DocumentSymbolRequest position
Summary:
I noticed that during the "Go to Symbol in Editor ..." request in VS Code the cursor jumps to the first character of the symbol's line.

This is contrary to what is happening in VS Code for TypeScript (which I'm using as a reference), where the same call lands to the beginning of the function name. I believe what is happening in TS is a nicer user experience as I probably want to be navigated to the function/class/etc. name that the symbol menu is showing me.

NOTE: This is my first diff in this project.

#thanks to mheiber for helping me with the setup to this project

Reviewed By: viratyosin

Differential Revision: D77998155

fbshipit-source-id: 9ad437d722431d2b02d8b6c6efe61d0e4a666b59