Expose the ability to get trait a const type is defined on via reflection
Summary:
Morally similar to D49778749. At some point we probably ought to just expose this across all the relevant reflection points, but that's a problem for a later date.

Currently it is impossible to go from "Cls::typeconst" to where the type const is actually defined when it is inherited from a trait (without using the file path, loading and parsing an AST, etc).
This exposes a convenient way of doing this.

Reviewed By: nt591

Differential Revision: D77872716

fbshipit-source-id: 59f27d26ea1dc9936a1dfcdc481058c3c31f7754