Break ReflectionFunction->invoke for top-level functions
Summary:
The way ReflectionFunction->invoke() works is, for non-closures, calling the `hphp_invoke` helper on the string name of the reflected function. This helper goes through the standard path for invoking functions, which reads `ForbidDynamicCallsToFunc` to decide what to do. When the level is 3, it throws an exception that the reflected function is being called without using a function pointer.

Instead of making it work by using HH\dynamic_fun and changing hphp_invoke's signature (takes a string) or ignoring the missing __DynamicallyCallable internally, this just deprecates the functionality from Reflection and deletes hphp_invoke.

Reviewed By: slinakm

Differential Revision: D78129818

fbshipit-source-id: b0bd407e464dc3d53c9f0481c607bf35b3d926b8