Add option to disable sending NewSessionTickets to Fizz server
Summary:
This change adds a command-line option `-no_session_tickets` to the Fizz server tool that allows disabling automatic sending of NewSessionTickets after a handshake is completed.

The Fizz server context already had built-in support for this functionality via the `setSendNewSessionTicket(false)` method, but there was no way to access this from the command line. This change exposes this functionality to users of the tool.

Reviewed By: ngoyal

Differential Revision: D78039425

fbshipit-source-id: 7c9649cb4aa59103ee4b1602cc316e33ffb4e093