Add onSinkConsumed in event handler and corresponding counter
Summary:
For sink, we are exposing two counters:

sink.on_next = total number of chunks received from client.
sink.chunks_consumed = total number of chunks consumed by application.

The difference is the total chunks saved in the queue.

There are one special case, when server side closed the sink, the those chunks will be discarded and won't be counted in consumed. But those should be minimum and won't impact the statistics.

Differential Revision: D78130042

fbshipit-source-id: 004430ae7a0cbf58afb48dca8702fbd2a836c100