Fix argument comment in mcrouter/FileDataProvider.cpp +1
Summary:
Argument comments ensure that the correct values are assigned to the correct arguments.

This diff is a manual fix (assisted by the script in D78190897) of one or more argument comments in mcrouter/FileDataProvider.cpp. Once existing violations are drawn down we will make mismatches between argument comments and argument names an error.

Reviewed By: dtolnay

Differential Revision: D78191367

fbshipit-source-id: d33aae9f42232f8ecd023507d60565bbf2e88adf