switch str internal type from bytes to str
Summary:
## TLDR:
- while this diff won't improve memory footprint (because the unicode string created from bytes keeps the byte array), it does improve runtime:
  - serialization: improved ~2%
  - deserialization: about ~1% worse
  - init from Python 1000 string list cinder benchmark: 50% improvement (2x faster) for immutable, 25% faster for mutable
  - init from Python 1000 string map benchmark: **~30% improvement for both immutable and mutable**
 - container field access:
   - 100-`str` immutable list iter **25% faster** (0.001181 ms-> 0.000881 ms)
   - 100-`str` mutable list iter **25% faster** (0.001932 ms -> 0.001543 ms)
   - 100-`str` immutable list idx **35% faster** (0.000811 ms -> 0.000530 ms)
   -  100-`str` mutable list **8% faster** (0.000390 ms -> 0.000360 ms ms)
- scalar field access: not much change, maybe 10% better. In other words, most of cost is caching. Unfortunately to preserve the error handling behavior below, we can't efficiently make it a primitive field.

Before in the init benchmark, thrift-python was  about 3x slower than py-deprecated at maps and 12x slower for lists, now we've closed gap to 6x for lists and 2x for maps.

## Implementation
We've known for a while that it feels wasteful for `string` fields to store two Python objects: the unicode `str` and the decoded `bytes` for serialization. The `PyUnicode_From` methods aren't much more expensive than `PyBytes_From`, so it's better to eagerly read the `string` type directly to unicode `str` rather than paying the cost of both Python C-API calls.

We don't even take a hit on serialization round-trip, because `PyUnicode_AsUTF8AndSize` stores an internal cached buffer with the encoded bytes, which we can use when serializing the object.

The tricky bit is handling `UnicodeDecodeError` from `PyUnicode_FromStringAndSize`. We can't use the normal cpp error handling mechanism , which captures the python error and bundles it in a gross `std::runtime_error`.

We need to keep the current exception API:
1. `deserialize` a struct with invalid unicode *does not* `raise`.
2. `serialize` a struct deserialized from invalid unicode *does not* `raise`. It produces the same bytes message that was deserialized.
3. Accessing a field with invalid unicode *does* `raise UnicodeError`.

That suggests we must keep `bytes` when the `UnicodeDecodeError` occurs during deserialize. Thus, the internal data type becomes `str | bytes`, with `str` representing the normal case, and `bytes` representing the original data that will produce deferred `UnicodeDecodeError` on field access.

Also note this involves updating the thrift-python C-API's `Extractor` type for `string` fields to use the `PyUnicode` methods as well. We use the `FallibleString` type to allow the `Extractor` and `Constructor` to keep the same behavior as serialize-deserialize.

Reviewed By: prakashgayasen, createdbysk

Differential Revision: D72875775

fbshipit-source-id: e4eb30123cf6b150853f7f2d29f722cd633d6913