Relativize internal loads to aid open source folly
Summary:
# Context

In the Open Source world we move folly up one directory (i.e. the folly folder is at the root, not a subfolder of `fbcode/`). Then, we model the folly project as the `root//` cell in OSS. This, of course, breaks all `deps` (and also the prelude and fbcode macros are different enough that they break too). To fix these problems we have a layer of macros that we wrote which "shim" `fbcode_macros` et al to what the prelude expects. E.g. update visibility, fix remap `external_deps`, and *critically* map the `deps` of the rules to point to the correct places (`fbcode//folly/X/Y:Z` becomes `root//X/Y:Z` and so forth).

Deps mapping cannot be done with loads, those are always cell-absolute so we need to figure out a way to handle these. We can MOSTLY ignore this problem because most `load()`s are loading common macros that we can just stick in the shims cell (and alias `fbcode_macros` et al to shims). However, there are the occasional project-specific loads (like `fbsource//xplat/folly:deps.bzl` here) which we need to handle somehow.

## This Diff

Relative loads presents a solution to this problem! Relative loads work no matter which directory the root is in, which makes sense. So we don't need to use the `fbcode//` or `fbsource//` cells (again, both point to the shims cell, which breaks the `load`) and instead can just let relative paths work in both places!

(this conveniently means that in the hgdirsync state that folly finds itself in the `deps.bzl` file will always be loaded from the correct cell!)

Reviewed By: mzlee

Differential Revision: D77186280

fbshipit-source-id: 689b2bc8544c050f2806f4414f5174a867110a93