Fix omitted $this initialisation checks
Summary:
We were shortcircuiting initialisation checks while determening which properties are initialised. This makes sense, however, the same logic that collects initialised properties also emits errors on premature `$this` usage. This leads to typeholes and compromises soundness.

This diff fixes it by preventing the shortcircuiting behaviour in various constructs.

The extebded support covers:
* `throw` arguments
* `for` condition, increment, and block
* `while` block
* `foreach` block

Reviewed By: geralt-encore

Differential Revision: D77967748

fbshipit-source-id: fdc07527aa91fb164b2abc5d4164d7706c4eb7e9