Accept non-splat lambda as splat function type parameter
Summary: The arity checking for lambdas is currently too conservative as it rejects a lambda that has concrete types for its parameters being passed to a function that makes use of a splat type in its function type parameter. Note that this only affects actual lambda values as arguments; lambdas assigned to locals are accepted already.

Reviewed By: madgen

Differential Revision: D78268460

fbshipit-source-id: b09f382c34fc632c262049561dcc1566bef9e169