log MONITOR_EXIT events
Summary:
I've noticed that we are not doing a great job tracking the life cycle of the monitor process:
When the server process dies with a non-zero exit code, we log a `TYPECHECKER_EXIT` event to Scuba. However, for the monitor we don't have a corresponding event we write to Scuba.

This diff adds an exit handler to the monitor, so that it logs `MONITOR_EXIT` to Scuba when the monitor crashes.

I'm interested in this, in order to track situations where the monitor goes down because of an exception.

Differential Revision: D78270315

fbshipit-source-id: ac1aabc98e8388c2ace8421333ab8404ca2056b9