Move typehole tests that allow types to lie into their own directory
Summary:
I've curated the tests in `typecheck/typehole` in order to

(1) Eliminate tests (or move into the main directory) for which the type hole bug has been fixed.
(2) In the case of multiple-instantiation unsoundness, add the enabling attribute `<<__UNSAFE_AllowMultipleInstantiations>>`
(3) Move "typehole-trusted" tests into a subdirectory so that they are called out specially.

Reviewed By: madgen

Differential Revision: D78404574

fbshipit-source-id: 1281a6b8fe45da0ef0572ce113489b9b006c6e0b