Remove "total" overload of SoftRealTimeExecutor::add
Summary:
The `SoftRealTimeExecutor` interface has an `add()` overload that accepts the number of times to run the function. This variant is not very useful, and in fact it is not used anywhere: it probably only exist because it was easy to implement it in `EDFThreadPoolExecutor`'s queue data structure.

On the other hand, it can be very hard to support with other queue implementations: it pretty much forces them to store the task in an object with shared ownerships, so that multiple threads can run the task at the same time. This is also error-prone for tasks that have mutable state.

Since nothing is using it, we can just remove it: if anyone ever needs to run a task multiple times, they can just add it multiple times to the executor. It is such a rare use case that it's not worth optimizing it.

Reviewed By: yfeldblum

Differential Revision: D78383752

fbshipit-source-id: 59618ff74d501f21149f86d003221d7e85b13378