add newtypes to built in pseudoconstants
Summary:
I was originally just going to do this for `__METHOD__` but while I was here I figured why not do all of them and see. Not sure if I need other hack/hhvm changes to make this viable?

What this enables is for a function, for example, to require that the type coming into it came from `__METHOD__` (via the newtype `BuiltinPseudoConstantMethod`) without breaking the world (since the newtype is still `as string`).

Reviewed By: viratyosin

Differential Revision: D41883768

fbshipit-source-id: a60f4dea5e11251f4d6ffec0629817430f893294