generate min/max static methods for thrift union tag enums
Summary:
In D77613228, we landed an optimization for `enum_find_name` which excludes union tag enums generated by thrift due to the missing of `min()/max()` methods in TEnumTraits. In this diff, we change the codegen to populate `min()/max()` methods for TEnumTraits<E> when E is a union tag enum.

The `enum_find_name` optimization will automatically apply to the union tag enums after this change.

Reviewed By: thedavekwon, ot

Differential Revision: D78344633

fbshipit-source-id: 6d4e26871104ec29661d353a22ab89b9221316b7