Type switch - add #t < T, #f < T
Summary:
The true and false tyvars are meant to represent parts of the scrutinee type, so it's valid to add the scrutinee type as an upper bound. Having this upper bound helps the typechecker draw conclusions in some cases and is a net improvement on errors in WWW after changes to use predicates for classes with erased generics.

The new errors in WWW are repro'd in P1870040373 -- I believe what's happening here is an incompleteness during the transitive closure check when adding bounds from the function call on the false-side tyvar from the inner check. The incompleteness seems to be related to how the new type splitting tends to split `null` and `nonnull` (and the latter is intersected with). A follow up change might be to improve this when we don't think the `null` splitting is likely to be useful.

Reviewed By: dlreeves

Differential Revision: D77958192

fbshipit-source-id: 6efda8713e73bc25ecdcadbc714b65f79fa5edbc