Report position of origin rather than upper bound for Axioms
Summary:
This is a different take on D76486783, but the core goal remains the same.

When we report a subtyping error based on projecting the upper bound of a given type, errors are suboptimal because we often point to a non-local position for the base type. This is due to some generic handling of the `Axiom` reason (along with Def and Proj) that flatly discards the original reason in favor of the constructed reason. This change flips that, reporting the original reason as the primary, and then elaborating about the bound's definition.

I tried a few variations of the wording e.g. "upper bound" vs "definition" and settled on this version where I carry back the bounded type's name for clarity.

One problem remains unaddressed from D76486783 -- I think the "But got " string should report the newtype rather than the bound, but at least the position and elaboration makes it clear. We're not throwing away information as mentioned in the discussion of the prior diff, where I'd construct a different prop with the outer types. It also preserves the more granular information from function type subtyping failures.

Reviewed By: enetsee

Differential Revision: D78459267

fbshipit-source-id: 761288946b98f74b73ea753a31296748579dacaf