Update user-facing hack errors to show flows from array accesses
Summary:
This change makes user-facing typing errors treat array access as a source of a local type, which makes it clearer to the user where the array's value type was computed (the subscript expression), as opposed to looking all the way through to the origin of the array's type.

in D78511642 I chose the span of `e1[e2]` as the witness, and that is reflected in the tests here. I considered highlighting just `e2`, but I felt it would be confusing having it point to the index.

Reviewed By: enetsee

Differential Revision: D78515650

fbshipit-source-id: 9a586ab92a24cabbe948d0209c5d4a6135c0716c