Typing_refinement.TyPartition.of_ty - produce predicates for generic classes with wildcards
Summary:
This diff is one component leading up to D75103426 which adds support for turning class hints with wildcards into predicates. These diffs will be squashed later since some of the individual parts use nocommits for parts that need to be filled-in in later diffs. They are separate for easier digestion.

This diff is the final diff and is what actually causes us to start using predicate-based logic for `is` conditions with hints with wildcards.

Most of this diff is for the errant case where a hint has too few generics in which case, we pad wildcards as necessary. Since wildcards use unique keys, we use the same mechanism as getting unique names for fresh generics, and so we must return the env (which is the rest of the changes in this diff).

I've also included some changes to fix some errors (mostly related to sound dynamic) that occur when this gets turned on.

I went through all the new WWW errors; the majority are "solved to nothing" errors or related to existing errors. There are a small number which are similar to existing incompletenesses related to sound dynamic.

Reviewed By: dlreeves

Differential Revision: D75103426

fbshipit-source-id: b4f1dd064f3b184c3e4d77c6b9d57ce6281c72f2