Typing_refinement.TyPredicate.to_ty takes an instantiation map
Summary:
This diff is one component leading up to D75103426 which adds support for turning class hints with wildcards into predicates. These diffs will be squashed later since some of the individual parts use nocommits for parts that need to be filled-in in later diffs. They are separate for easier digestion.

This diff updates `Typing_refinement.TyPredicate.to_ty` to take a map from wildcard key strings to locl_tys. This map is meant to come from `Typing_refinement.TyPredicate.instantiate_wildcards_for_predicate` which is to be called from `typing.ml`; for now, we use nocommits to mark where we're currently passing an empty map

For other current callers, a new function `to_ty_without_instantiation_opt` is added that does not require this map, but returns `None` when the give predicate has Wildcards

Reviewed By: dlreeves

Differential Revision: D75103427

fbshipit-source-id: 1e67f49afa551758bb76d56b9d5621e02b22a35e