Typing_refinement.TyPredicate.instantiate_wildcards_for_predicate
Summary:
This diff is one component leading up to D75103426 which adds support for turning class hints with wildcards into predicates. These diffs will be squashed later since some of the individual parts use nocommits for parts that need to be filled-in in later diffs. They are separate for easier digestion.

In this diff we add a new function that, given a predicate, instantiates new types for the wildcards and returns a map from the wildcards in the predicate to information about the new types. This map will be used for two purposes:

* Converting the predicate into a locl_ty using the map as a dictionary for how to convert the Wildcards in ClassTags
* Simplifying the environment using `Type_parameter_env_ops.simplify_tpenv`

Reviewed By: dlreeves

Differential Revision: D75103431

fbshipit-source-id: cb5ed324ccb2321b92733387451930bdfd9f65c6