Typing_refinement - produce "uninstantiated" propositions
Summary:
This diff is one component leading up to D75103426 which adds support for turning class hints with wildcards into predicates. These diffs will be squashed later since some of the individual parts use nocommits for parts that need to be filled-in in later diffs. They are separate for easier digestion.

As part of partition_ty, we produce propositions which are to be used to update the environment in true and false branches. These propositions involve types from the predicate. But when we have a predicate with wildcards, we cannot turn the predicate into a type without instantiating the wildcards which we want to be careful about when we do this.

This diff changes partition_ty to not produce subtype propositions but instead "uninstantiated" propositions that can later be instantiated by passing in an instantiation map (that we expect to create in typing.ml)

This allows us to decouple partitioning from instantiation of wildcards which is helpful since not all uses of partitioning makes use of these propositions and doesn't need to do instantiation.

Reviewed By: dlreeves

Differential Revision: D75103425

fbshipit-source-id: d9256ac743856cad9e3de145e48077dfa28614e7