ClassTag of Ast_defs.id_ * type_tag_generic list
Summary:
This diff is one component leading up to D75103426 which adds support for turning class hints with wildcards into predicates. These diffs will be squashed later since some of the individual parts use nocommits for parts that need to be filled-in in later diffs. They are separate for easier digestion.

This diff updates the `ClassTag` variant to take a list of `type_tag_generic` instead of a list of `locl_ty` where `type_tag_generic` can be one of:

* `Filled of locl_ty` - for use when a generic is provided (e.g. reified generics)
* `Wildcard of string` - for use when a wildcard is provided -- this is analogous to an existential type variable -- the string serves as a way of distinguishing the variables when we instantiate types for the wildcards and then use those types to turn the predicate into a type.

Reviewed By: dlreeves

Differential Revision: D75103428

fbshipit-source-id: 1f5051629d097f593e15783d0a585055be503530