Typing.Expr.type_switch - instantiate wildcards
Summary:
This diff is one component leading up to D75103426 which adds support for turning class hints with wildcards into predicates. These diffs will be squashed later since some of the individual parts use nocommits for parts that need to be filled-in in later diffs. They are separate for easier digestion.

This diff makes use of all the mechanisms built up in previous diffs to:

* produce an instantiation map for predicates with wildcards
* use this map when producing types and updating the environment for the true and false branches

Reviewed By: dlreeves

Differential Revision: D75103430

fbshipit-source-id: 7d3ed1ef397e64eca1e7e9a47468785e6c0ae701