Add sampling config for CheckedUnsafeCast
Summary:
Introduce a config `Eval.CheckedUnsafeCastSampleRate` that controls the rate at which the instruction (`VerifyTypeTS`) is interpreted as a type-check at runtime.

If set to 0, never check the types: for JIT, optimize the bytecode away completely.

If set to 1, always check the type, and log a notice if the check fails.

If set to n, check at a sampling rate of 1 in n.

If `Eval.CheckedUnsafeCast=false`, in the JIT, don't generate code, and likewise for repo-mode in HHBBC.

So to "stage" things, you can produce a repo with `Eval.CheckedUnsafeCast=true` and then when executing the repo, you can set the sampling rate (see in Test Plan below).

Reviewed By: vassilmladenov

Differential Revision: D77233760

fbshipit-source-id: 83095b7bd162deb903a491217c999325047957a7