Fix type widening bug in JIT
Summary:
This is another attempt at D76878556. In the previous fix it was possible for an `AssertType` to appropriately narrow a type during region formation and then later be removed during simplification due to non-monotonicity.

This diff introduces a `refine()` method for `Type` that can be used to ensure when choosing between RATArray specializations the refinement is always selected. This is further evidence that we really need to revisit the way RATArray types are handled by the JIT as this type of non-monotonic type relationship is likely to cause other problems.

Reviewed By: binliu19

Differential Revision: D78590911

fbshipit-source-id: afb49e9ad44609c1a1a89cdf6bcf8692ca40a0a4