make throttle_time_ms configurable
Summary:
This diff adds the option `edenfs_file_watcher_throttle_time_ms` to `ServerLocalConfig`, with a JK overrride in place.

It sets the value of the of the `throttle_time_ms` parameter passed to Eden's stream_changes_since API. This means that this is the minimum period (in milliseconds) between each time Eden will send us a change notification.

Reviewed By: madgen

Differential Revision: D78818059

fbshipit-source-id: e427cc70b014241c05f39971cdfe968e9ddce1ca