Raise required CMake version to 3.18
CMake 3.31 deprecates compatibility with CMake < 3.10 and CMake 4.0
removes compatibility with CMake < 3.5 entirely. As the required CMake
version hasn't been raised in 12 years, select 3.18 as a reasonably recent
minimum CMake version with wide availability (Ubuntu Jammy ships 3.22,
Debian oldstable Bullseye ships 3.18). Remove cmake_policy() calls that
now default to NEW due to the higher minimum version requirement.

A further benefit of this is that CMP0065[1] now also defaults to NEW,
which means that the `hhvm` binary is no longer linked with
`--export-dynamic`, making the (stripped) binary around 30MiB smaller.
OSS could also choose to supply an explicit dynamic list as a followup
as done in D2747334 and D67114451, but it doesn't seem to be required
just now.

[1] https://cmake.org/cmake/help/latest/policy/CMP0065.html