factor out concepts shared between stream and sink
Summary:
Move a couple of concepts out of `lib/python/server/util.h` (bad naming) and put them in `lib/python/streaming`. This is a place to put concepts shared between stream and sink.

In particular, the `PythonStreamElementEncoder` that is used by stream on the server side, and will also be used by sink on the client side.

Reviewed By: yfeldblum

Differential Revision: D79021539

fbshipit-source-id: 53fa8dd139ad38e6c5e714d7f09a42cba49c1e95