Substitute lower or upper bounds of type params appearing only co- or contravariantly
Summary:
In some cases a polymorphic function type can be written as an equivalent monomorphic type. Typically we warn users about these cases ('redundant generics') when written in function declarations but the current typing of function pointers / meth_caller will produce redundant generics when it quantifies over class level  generics which only appear co- or contravariantly in the resulting function type.

Although the types are correct, we prefer to generate minimally-polymorphic function types because (1) typing is more performant and (2) they are easier for users to understand.

This diff adds an additional pass to the generated function types which ensure any generics occurring only co- or contravariantly *and* not appearing as the rhs of class refinement are 'inlined' i.e. replaced by the intersection of their upper bounds or union of their lower bounds

Reviewed By: mheiber

Differential Revision: D78662593

fbshipit-source-id: 4cbc2f30f09806b85018dba781bcce2e2afe1bb6