log build drift to scuba
Summary:
Add the compiler option `--log-package-build-drift`.  If set to true when building the repo with PackageV2 U SymbolRefs, it logs to the new Scuba `packagev2_www_build_drift` table (owned by Hack) the files with are pulled by PackageV2 or SymbolRefs but not both.  This will this enables monitoring in (almost) real time:

  - which files are pulled in by Packages but not SymbolRefs (and are thus candidates to be removed from the static 'prod' package), eg. https://fburl.com/scuba/packagev2_www_build_drift/x67p6pud

- which files are pulled by SymbolRefs but not Packages (this in not empty right now, but once we turn on static enforcement, any file in this category will denote a Hack Package boundary bug):, eg: https://fburl.com/scuba/packagev2_www_build_drift/8rx0p12g

Reviewed By: mdko

Differential Revision: D69188466

fbshipit-source-id: 497ddf761add33dce4433da5c0d8d2217d911755