Don't write translation sizes to the TC
Summary: Instead of writing sizes of cold and frozen translations directly in the TC, we can store it outside to save space to JIT more. Sometimes theres no cold or frozen code being emitted at all, but we'd still write sizes of `0` in the TC which meant `16+16` bytes are used each to do alignment.

Reviewed By: ottoni

Differential Revision: D79108091

fbshipit-source-id: 8506d001a7e2b9f296d949c5cd47123de717e33a