guard all calls to getPackageInfo
Summary:
HHVM fatals when invoked via `hphpd -h` because in this setting `getRepoOptionsForRequest()` in `getPackageInfo()` returns `null`.  This diff guards all the invocations of `getPackageInfo()` to the case where an ActiveDeployment is set.

Remark: an alternative is to get `getPackageInfo()` to fail gracefully, but I believe that guarding all runtime accesses to PackageV1 code is valuable.

Reviewed By: sinancepel

Differential Revision: D78006232

fbshipit-source-id: d5d76add4107fafbbac0bd386ccff54cc6954764