Delete breakpoint capability and enable server mode for notebooks
Summary:
## Context
Fbdbg is an unowned and unmaintained Hack debugger that's used in Hack Bento Notebooks and on the command line as a repl.

Although fbdbg enables debugging capabilities, it's rarely used among Hack programmers, with the VSCode debugging extension being preferred.

Deprecating fbdbg allows us to
1. Reduce support burden on the HHVM team
2. Enable more capabilities for Hack notebooks easily

["Deprecating fbdbg" Google Doc](https://docs.google.com/document/d/12CyH78_LT6CbXMQF603-_75DwkVlpmYo2MW9bfEFMM4/edit?tab=t.0)

## This diff
This diff removes the ability to add breakpoints to HPHPD. I want to perserve some breakpoint-like functionality for Hack notebooks (e.g. while interrupting), which is why everything hasn't been deleted yet. I also deleted a couple of tests that were only testing breakpoint functionality and updated the rest; to be honest, I could probably be deleting more tests.
The one-liner in `program-functions.cpp` is to enable connecting to the server on notebooks.

Note: The real change in this diff is all in `fbcode/hphp/runtime/debugger/`. There was a ton of expect files I had to update in this diff that you can review if you would like!

Reviewed By: nt591

Differential Revision: D78750746

fbshipit-source-id: 2613646d29a32d8bba2be218b848cff530b08582