test shows OmniClient sink upload works
Summary:
This is a test validating OmniClient sink functionality using C++ server.

Note the real Python sink client serializes payloads before handing them to the OmniClient as IOBuf, and the Python server deserialize IOBuf back to the private RPC types before giving them to the handler.

But in this test, the OmniClient is serializing, but the C++ server doesn't have the deserialize logic, so we have to manually simulate the deserialization inside the handler. Hence, the service method type of the `sink` is `IOBuf` even though it represents a `sink<StreamChunk, std::string>`.

The next step is adding coverage of exception handling.

Reviewed By: prakashgayasen

Differential Revision: D79405824

fbshipit-source-id: c2322cb015979cb1920d1d2bb6e3bcd44d3d685a