Update verify.py so it can update snapshots
Summary:
## What

We update our verify script to automatically update snapshot tests when the UPDATE environment variable is set to 'always'.

UPDATE=always is currently used by review.sh, the script that updates snapshots.

> The verify script is used internally by our buck snapshot (.exp) tests.

## Why

The next diff in the stack will modify our buck tests to propagate the UPDATE setting, see that diff for more context.

## How

Extract out the logic that we were using to say how to update snapshots so that we can use it from two places:
- (existing behavior) to give the instructions on how to update the snapshots
- (new behavior) To actually execute the commands to update tests

Differential Revision: D79342781

fbshipit-source-id: 1038406e7b86621ba4653911703d9fbaba2c81d3