Introduce variable_component AST abstraction for variable lookups
Summary:
Introduces a new Whisker AST abstraction `variable_component`, which comprises two components - an optional explicit prototype specifier, and a required identifier.

This introduces the AST node with no change in behaviour.

The current parsing behaviour is unchanged - the prototype component will always be initialized as `nullopt`.

The template execution behaviour is also unchanged - only the component name is used, while the prototype specifier is not referenced.

Further changes will introduce usage of the prototype specifier in variable evaluation, and parsing to extract the prototype specifier from template source.

Reviewed By: praihan

Differential Revision: D79405311

fbshipit-source-id: 9e86996dc720676792d0720a47a0658293f8d196