Update use of GetChildAtIndex to reflect changes in API for next version of LLDB
Summary:
The API for GetChildAtIndex was changed in https://github.com/llvm/llvm-project/pull/140065. So that "True" at the end means something different now ("treat_as_array" rather than "can_create_synthetic"). This change will be present in the next version of LLDB used internally.

Happily it looks like there's a way to update things that both doesn't break the tests with the current version while updating it for the next version; otherwise we'd have to use something like [this](
https://www.internalfb.com/code/fbsource/[959f6221198f]/fbcode/sand/tests/lldb_test_case.py?lines=53) in order to gate things (leaving this as a note for future changes).

Reviewed By: zhyty

Differential Revision: D79487113

fbshipit-source-id: 946f0f6d8fee72cfaf85c710fb613dda89003c6f