add timeouts to expensive operations
Summary:
This diff adds timeouts to potentially expensive operations in `Edenfs_watcher`.  The timeout is configurable as a new config option `timeout_secs`.

Concretely, we add timeouts
1. whenever we talk to Eden or Hg directly
2. whenever we try to acquire the lock to the shared state. This is mostly aimed at detecting deadlocks. I switched our locks from `std::sync::Mutex` to `parking_lot::Mutex`, since that mutex type has a `try_lock_for` function using a timeout value.

The motivation for 1. is that if these interactions with external systems hang indefinitely, we currently have no way of detecting this, except for hh not processing file changes or not answering client requests. The motivation for 2. is similar, in the unlikely case that there are deadlocks in the code it would be good to have some way of finding out.

I've picked a value of 60 seconds for the timeout. My intention is for this to be pretty generous. My intuition is that If any interaction with hg or Eden takes longer than this, that's an indication that something is wrong.

Reviewed By: madgen

Differential Revision: D79294084

fbshipit-source-id: 340be9d90613de7b8a91605dd19507fd81d6d546