new option edenfs_file_watcher_sync_queries_obey_deferral
Summary:
This diff adds a config option `edenfs_file_watcher_sync_queries_obey_deferral`. It determines how `ServerNotifier.get_changes_sync` acts when `Edenfs_watcher` is used and we are currently in a deferred state (e.g., because meerkat is running).

If the option is `true`, `get_changes_sync` will not return file changes while we are deferring. If `false`, `ServerNotifier.get_changes_sync` will act like it does when using Watchman: It returns all available file changes, no matter if we are currently in a deferred state or not.

The motivation is to allow`Edenfs_watcher` to behave more like Watchman, to investigate some divergences in our metrics between the two.

In order to make `edenfs_file_watcher_sync_queries_obey_deferral=false` work as intended, I also needed to enable us to break out of the `pump_async_updates` loop in `ServerMain.query_notifier` if there is a notfication, but we don't actually get changes from `ServerNotifier.get_changes_sync`.

Reviewed By: madgen

Differential Revision: D79352047

fbshipit-source-id: 0a9e0eccc33bcd385e4b298d4de3fa2d329a102c