Symbolizer check ELF executable segments for instruction addr
Summary:
Fix bug in `folly::symbolizer::Symbolizer#symbolize` that would map an instruction to the wrong ELF file and cause the instruction to be labeled with `(unknown)` by `folly::symbolizer::StringSymbolizePrinter#println`.

The bug occurred when an instruction addresses was in range of an ELF section that would [NOT appear in the memory image of a process](https://man7.org/linux/man-pages/man5/elf.5.html#:~:text=If%20this%20section%20appears%20in%20the%20memory%20image%20of%20a%20process%2C%0A%20%20%20%20%20%20%20%20%20%20%20%20%20%20this%20member%20holds%20the%20address%20at%20which%20the%20section%27s%20first%0A%20%20%20%20%20%20%20%20%20%20%20%20%20%20byte%20should%20reside.%20%20Otherwise%2C%20the%20member%20contains%20zero.) (i.e., `sh_addr = 0`). For example, this happens in  `libstdc++.so.6` that has a section with `sh_addr = 0` and `sh_size = 6795731`. The bug is normally hidden because `libstdc++.so.6` is loaded after other `.so` files, so the address is found before checking `libstdc++.so.6`. It occurred with `fbcode//mode/dev-nosan` because that mode creates an executable that loads `libstdc++.so.6` before the other `.so` files. The `Symbolizer#symbolize` algorithm joins addresses to elf files with a double `for` loop over the dynamic linking library's [link_map](https://man7.org/linux/man-pages/man3/dladdr.3.html#:~:text=link.h%3E%0A%20%20%20%20%20%20%20%20%20%20%20%20%20%20as%3A-,struct%20link_map,-%7B%0A%20%20%20%20%20%20%20%20%20%20%20%20%20%20%20%20%20%20%20%20%20%20ElfW(Addr)%20l_addr) and the `FrameArray` obtained from libunwind.

The fix is to check if the instruction is contained inside of the ELF file's segments (program headers) that are executable and loaded.  Segments are the loader's view and sections are the linker's view.

An alternative change to fix the tests is to filter out sections with `sh_addr == 0`.

Reviewed By: Jlalond, rafaelauler

Differential Revision: D79476989

fbshipit-source-id: ba167677c89e8fbf1f1ca5ecc474955e29556232