Make xoshiro256pp vector type a template argument
Summary:
The AVX2 version of xoshirp256pp generates a different sequence of numbers given the same seed. This is a (hacky) fix to give users control over which variant is used with a template argument.

The default template value should ensure there is no change to which variant is used, but the variant used by default will remain the same (AVX2 if available).

Reviewed By: Orvid

Differential Revision: D79451125

fbshipit-source-id: 7e9bb786e5309e535794c7c73f6eb2b8030cb2d8