implement ClientSink cython
Summary:
Now that the code is in a more re-usable state, we can implement the cython.

Use a generated python sink element handler method because that has to be able to catch more than one element type.

For the final response handler, we can do that cleanly because either `success` is set on the returned struct, or one of the exception fields is set. We can just iterate through and raise the non-None one without any gencode at all.

Reviewed By: prakashgayasen

Differential Revision: D79485622

fbshipit-source-id: 3cbd2b4ca0d8d37cf2ac1cd43e6ce82c12c4abd9