Support class pointers in enum classes
Summary:
This was an oversight by me -- this function is never called on regular enums and the comment was wrong. Implementation is for legacy enums `extends Enum` and enum classes, which is why `enum E: class<C>` never raised the error observed on enum classes (see V1).

Some more cleanup is necessary to truly finish this, but for now this fixes enum class's intentional support for class pointers.

Reviewed By: viratyosin

Differential Revision: D79917193

fbshipit-source-id: d9e2b679dcaf19983eb324c894c187f21b0daff9