Decouple serialization and types more
Summary:
We are decoupling serialization and types to speed up modular builds. This is based on the observation that serialization code consists of large templates that few users need to consume directly. Decoupling is mostly complete on the target level and the remaining part is making the `<module>-cpp2-serialization` target a private dependency of `<module>-cpp2-types` as in the following diagram (currently it is a public dependency):

```
          [<module>-cpp2-types-headers]
                  ^          ^
          (public)|          |(public)
                  |          |
[<module>-cpp2-types] --> [<module>-cpp2-serialization]
                   (private)
```

The main remaining blocker is that the generated `<module>_types.cpp` includes `<module>_types.tcc` which is a header currently combining some implementation details of types and serialization code. It also contains some remnants of serialization code itself even after most of it has been moved to separate translation units.

This diff decouples serialization from types by moving the serialization code that is currently in `<module>_types.tcc` to `<module>_custom_protocol.h` and the one that is currently in `<module>_types.cpp` to `<module>_types_serialization.cpp`. `<module>_custom_protocol.h` headers are arguably a better fit for this than `<module>_types.tcc` which were often misused (D79115339).

This also prevents users from creating semi-broken targets with some headers not includable due to long paths on Windows (see e.g. D79198470).

And last but not least it slightly improves build speed as measured by `benchmark_diff --diff D79106423 --top-slow 1000`:

```
+-----------------------------------------------------------------+------------+--------------+----------------------+--------------------+
| Measurement name                                                | End date   |   All fbcode |   Monetization Users |   Monetization Org |
| https://www.internalfb.com/phabricator/paste/view/P1898352612   |            |      avg EYS |              avg EYS |            avg EYS |
+=================================================================+============+==============+======================+====================+
| benchmark/D79106423/1754691520_a2f9eb1c                         | 2025-08-08 |      0.01398 |              0.02013 |            0.01664 |
+-----------------------------------------------------------------+------------+--------------+----------------------+--------------------+
```
Per week measurement data in markdown:
[07/11-07/18](https://www.internalfb.com/phabricator/paste/view/P1898344930) - -0.00068 EYS - 1875 builds - Top user: nikitabhalla: 0.00307 (-451.64%)
[07/18-07/25](https://www.internalfb.com/phabricator/paste/view/P1898350876) - 0.04267 EYS - 1724 builds - Top user: bzan: 0.00317 (7.44%)
[07/25-08/01](https://www.internalfb.com/phabricator/paste/view/P1898348469) - 0.00424 EYS - 1713 builds - Top user: chenyh: 0.00329 (77.74%)
[08/01-08/08](https://www.internalfb.com/phabricator/paste/view/P1898352431) - 0.00969 EYS - 2058 builds - Top user: neelansh: 0.00511 (52.72%)

More context: T206890027.

Reviewed By: iahs

Differential Revision: D79106423

fbshipit-source-id: 54897209026837371f5ceadf6d9ab0f41d466cf8