Add get_newtype_super helper function
Summary:
[This is a second attempt to eliminate the inline bound on `Tnewtype` in favour of a helper call `get_newtype_super`. This time, we will introduce uses of the helper in stages, in order to avoid perf regression. The original diff was D76805652.]

The `Tnewtype` constructor for `newtype` uses contains inline in the type its bound i.e. the as type from the definition instantiated with its type arguments. This contrasts with with practice elsewhere (bounds on abstract type constants, bounds on generics, inheritance for classes, etc) where we compute the bounds on demand. When creating a `Tnewtype` we must be careful that the bound actually matches the definition, and in general it's a potential source of bugs. Also, it doesn't easily extend to multiple bounds.

Instead, we will look up the bound on demand, using the `get_newtype_super` helper function. In a future diff, we will extend the "type-directed" interpretation of `TypeStructure<t>` (currently limited to direct shape indexing) by simply returning a bound from `get_newtype_super` that is determined by `t` .

This diff is the first in a stack: it simply introduces the helper function, but does not make use of it.

Reviewed By: vassilmladenov

Differential Revision: D79888539

fbshipit-source-id: e29590a7926dd3545f7f6a56aac35a5b6f96f8ae