generate handler methods
Summary:
The ClientSink.sink method executes outside the client / server context with user async generator as input. Since it's allowed to `raise` more than one exception type, we need to generate a python handler method to catch and translate these appropriately. If feels cleaner to do this as a `classmethod` of the `sink_elem` rpc type than make it a random method we have to generate in the client and somehow inject into the stream, which would mess up the relatively simple structure.

With respect to the final response type, we don't have to do this because the exception is set as a field, so we can iterate the fields in a type erased way from cython lib without generating any python!

Reviewed By: prakashgayasen

Differential Revision: D79227684

fbshipit-source-id: 2b1288a4b45e6f8a3d563af24066fc4b8190c6a1