Delete unused debugger commands
Summary:
## Context
Fbdbg is an unowned and unmaintained Hack debugger that's used in Hack Bento Notebooks and on the command line as a repl.

Although fbdbg enables debugging capabilities, it's rarely used among Hack programmers, with the VSCode debugging extension being preferred.

Deprecating fbdbg allows us to
1. Reduce support burden on the HHVM team
2. Enable more capabilities for Hack notebooks easily

["Deprecating fbdbg" Google Doc](https://docs.google.com/document/d/12CyH78_LT6CbXMQF603-_75DwkVlpmYo2MW9bfEFMM4/edit?tab=t.0)

## This diff
This diff removes the following commands from `hphpd` since we no longer support CLI debugging:
* Down (a frame)
* Up (a frame)
* Next
* Frame
* Step
* Out

I kept CommandWhere and CommandVariable since we might need it in the Hack Notebook. We definitely use CommandVariable (which outputs variables), and CommandWhere could be useful for interrupting in Hack Notebooks.

These are the most useful files to look at:
* fbcode/hphp/runtime/debugger/debugger_command.cpp
* fbcode/hphp/runtime/debugger/debugger_client.cpp
* fbcode/hphp/runtime/debugger/cmd/cmd_help.cpp

The other files are deleted or updated tests.

Reviewed By: nt591

Differential Revision: D80022433

fbshipit-source-id: 955a73f8e596f903efdb110ee01a8a762099e197