Add warning for e1 ?? e2 for always and never null cases
Summary: This extends the Is_as_warning tast check for `if ($foo is null)` etc, to cover binops `e1 ?? e2`, by analogy to an expansion `e1 is null ? e2 : e1`. We exclude cases where e1 is an array subscript expression, as that conflates key existence with the value being null.

Reviewed By: viratyosin

Differential Revision: D79935307

fbshipit-source-id: 5642cd679cad90b4f29c1efc7300a76d8fb4a737