Thrift Compiler: Generate ServiceTraits.
Summary:
ServiceTraits identify the set of types that are used in an RPC interface.
In an interceptor context, the set of types in ServiceTraits matches all of the
potential types that may be associated with a parameter, return value or exception
presented by interceptors as a TypeErasedRef.

This will be used in subsequent diffs to provide better visitation mechanisms
for use in interceptors.

Reviewed By: praihan

Differential Revision: D77686322

fbshipit-source-id: abc2eecc70fcc77d71671ade3621e9b2dad2fb7e