impl split logic
Summary:
This diff groups the types in a thrift file (typedefs, structs, enums) into different chunks. Types that are dependent on other types are always in the first chunk. The rest of the types are even divided into the remaining buckets.

These new functions and variables are used in later diffs in this stack to divide types across multiple crates to parallelize the build process.

Reviewed By: dtolnay

Differential Revision: D79600672

fbshipit-source-id: 3db8613bc14bbe42abaecfb6bd511877a767ab27