Removing usage of deprecated t_type:: get_type_value
Summary:
In order to deprecate the t_type:: get_type_value and t_type::get_type_id , we need to remove the usages of these methods.
This diff is to replace the use of these methods wherever possible with alternate implementation.

This diff is not a full deprecation, it is only removing refs from compiler code

a.get_type_value() === expected value  replaced with a.is<t_*>
- before -> if(a.get_type_value() == t_type::type::enum)
- now -> if(a.is<t_enum>)

tContainer::get_type_value() replaced with tContainer::containerType()

Reviewed By: iahs, vitaut

Differential Revision: D79701843

fbshipit-source-id: 289d7539b763e69e6c98453ec8ceeb8693846725