do not modify zero copy receive IOBufs in Thrift parser
Summary:
FrameLengthParserStrategy by default will attempt to optimise the IOBuf passed up in readBufferAvailable(). But this is not desired for IOBufs from IoUringZeroCopyBufferPool - which should not be destructively modified in userspace without being cloned first. The optimisations causes IOBufs to no longer be page aligned, or for IOBufs to have a length of 0.

Fix by marking IOBufs from IoUringZeroCopyBufferPool as externally shared. This prevents IOBufQueue from modifying the IOBuf.

Reviewed By: ot

Differential Revision: D79510459

fbshipit-source-id: b220562a86d3028befdfd9ae8c23faed1c02497f