Add HH\type_structure_class migration helper
Summary:
`HH\type_structure_classname` is an (unexposed) helper meant to aid in the calculation of `type_structure(C::class, 'T')['classname']`, and is also explicitly reduced to in HHBBC

https://www.internalfb.com/code/fbsource/[d003b9ca547a972a1475ba2aeb7ec29c1e3d2374]/fbcode/hphp/hhbbc/interp-minstr.cpp?lines=2266-2267

This creates an adjacent helper that returns the pointed class, as an alternative to looking up the classname field when fetching the type structure from a dynamic class. Unlike the classname variant, I haven't bothered to support the type alias case (second argument optional). This can be added later if necesssary.

Reviewed By: ricklavoie, jano

Differential Revision: D80067535

fbshipit-source-id: 0549348228ade63ebb074cf3a70405fb712c2dd3