TypeRefVisitor: Visitation for objects identified by TypeErasedRef
Summary:
The motivating problem for this is that Thrift interceptors present
a TypeErasedRef to represent arguments/return values to RPC methods
being intercepted.  We want to be able to provide a function that
consumes some type that may be present in an RPC payload. Now, how
does one find all of the potential instances of that type in the
RPC payload that is represented by the TypeErasedRef?

A visitation is necessary to walk through the type structure of
the RPC payload. The TypeErasedRef does not contain the compile-type type
information needed to instantiate the visitation templates.

The solution provided by this change is to provide a mechanism
to build a "VisitorMap", which maintains a map of all possible visitors
for the payload types associated with a Thrift RPC service.

The VisitorMap is used by a "visitTypeRef" function that given
a TypeErasedRef (presumably of a payload of the RPC service) uses a VisitorMap
to drive the invocation of a visitor function that consumes some type
that may be found within the payload.

The specific use case for this is to be able to identify all "Tensor"
types that may be found within an RPC payload (so as to potentially
offload the transfer of this data to a bulk transport).

Reviewed By: praihan

Differential Revision: D77686320

fbshipit-source-id: d335bc9dd3b13578a28542f7910520a6da4106f9