Thrift: Interceptor support for response inspection.
Summary:
Presents RPC responses to interceptors.

Library code trivially adds the necessary paths to present RPC
results to interceptors.

The trickier part is that the client stub generated code must be modified.
Prior to this, response interceptors are called too early.  This change
adapts code generation templates to defer the interceptor invocation
until after deserialization, and to correctly handle the success/result
vs exception paths.

This alone has some subtantial syntactical complexity in the mustache
code because there are multiple cases that must be coded for independently.
- Stream?
- Interaction?
- Oneway?
- void return?
- Combinations of the above.

Additional complexity comes in from the face that the best model for the
code is to use "recv_wrapped_" to obtain the result values.  But this
fails for streams, as it would require (a non-existent) default ctor
for stream objects.  And thus these paths must rely on a less optimal
folly::makeTryWith approach.

Variants of processClientInterceptorsOnResponse are provided, accepting
different forms of result/exception.  Doing this simplifies the .mustache
templates - adapting the C++ code to the constraints of the template vs
doing type conversion/construction in template code.

Reviewed By: tlj77

Differential Revision: D77686325

fbshipit-source-id: 36389e4631b4084f44f910da7adb720599e48d63