stream elements: prevent exception field name collisions
Summary:
Take a fix we applied for regular service exception fields, and apply it to stream element exception fields. This prevents bugs due to users re-using the same exception field name.

__From original diff__
We're relying on user input (`throws` "field" names) for internal variable names. Specifically, we're using these for the generated per-function response `struct`. There's a couple ways for this to be buggy:
1. If an exception is called `"success"`, then it will be impossible to access a non-exception return value.
2. If two exception fields have the same name (as in recent user post), then it will be impossible to access the first error if returned.

The solution is to mangle the internal field name with the field index to ensure there's no chance of collision (leveraging exsting IDL validation). Can't use field key (id) because it may be negative

Changing field names is wire compatible, so it should be ok to push this without explicit compatibility code.

Reviewed By: prakashgayasen

Differential Revision: D80229104

fbshipit-source-id: 94f98ae569a59113ce9156205eed699e5ba5af09