Make RocketClient control flow resemble server-side equivalent
Summary:
BiDi will make some currently terminating methods conditionally terminating, so the existing method of returning a StreamChannelStatus to tell the caller whether to call freeStream will become unwieldy. In addition, the dissimilarity with the equivalent server code increases the overall complexity of this codebase.
Make callees responsible for calling freeStream, and use the same bool convention as the rest of the streaming code. This increases safety by eliminating the need to separately track whether the callback object is destroyed.

Reviewed By: tlj77

Differential Revision: D79916934

fbshipit-source-id: bc9c066355b48a75b73240ca4bc45ce288f396cc