Improve indentation in generated Rust files
Summary:
When a `{{> ...}}` is indented like this:

https://www.internalfb.com/code/fbsource/[0912debc38df6a8b7c662364418d685d99fef243]/xplat/thrift/compiler/generate/templates/rust/types.rs.mustache?lines=74-76

it ends up producing output that is also indented, despite that indentation being undesirable in the output code.

https://www.internalfb.com/code/fbsource/[d90ea9de7af59f966bb8bdb1c78301984d4eef4e]/xplat/thrift/compiler/test/fixtures/basic/out/rust/gen-rust/types.rs?lines=14-18

On the other hand, just erasing indentation from our templates makes the templates unreadable.

This diff erases the indentation but also inserts some '**=========**' separators to maintain some hierarchical structure in the templates.

Reviewed By: dtolnay

Differential Revision: D80294569

fbshipit-source-id: 1df52dcbeaaf2312d0cfbf96ef1e5b92de9aff58