Move `contains` to folly as a utility
Summary: Add `folly::contains`, which, given some container, will pick the right `contains` function to use (e.g. `contains`, `find` or `std::find`), making it more convenient to work with template types. As a side-effect, users can also change a type from, say, a vector to a set, without a bunch of extra changes. Overall, it's a nice to have, and is not present in the standard library or folly.

Reviewed By: Gownta

Differential Revision: D79346399

fbshipit-source-id: 26c20cb1d9957d2e3d0d06029c006a466f585315