Add NoDisjointUnion attribute to idx and idx_readonly
Summary: Code such as `idx($v, C::class)` for `$v:vec<int>` is currently permitted by Hack because the key type on `KeyedContainer` is covariant. But we can use `NoDisjointUnion` to warn whenever Hack infers a disjoint union type for the key, usually an indicator that there is an error. This will also pick up cases such as `$v[1]` for `$v:vec<bool> | dict<string,bool>` and although this is valid (i.e. it doesn't always fail at runtime) it's often a bad smell.

Reviewed By: madgen

Differential Revision: D80333535

fbshipit-source-id: 692137ca2d37c9eaa77d10c509a03a7d920a69a0