Move serialization targets to private dependencies
Summary:
We are decoupling serialization and types to speed up modular builds. This is based on the observation that serialization code consists of large templates that few users need to consume directly. Decoupling is mostly complete on the target and source level and this diff implements the final and most important bit: making the `<module>-cpp2-serialization` target a private dependency of `<module>-cpp2-types` as in the following diagram (previously it was a public dependency):

```
           [<module>-cpp2-types-core]
                  ^          ^
          (public)|          |(public)
                  |          |
[<module>-cpp2-types] --> [<module>-cpp2-serialization]
                   (private)
```

Export serialization libraries as header units.

Rename `<module>-cpp2-types-headers` to `<module>-cpp2-types-core` since these targets no longer contain just headers but all types code that doesn't depend on serialization. Previously all code was in `<module>-cpp2-serialization` which was a bit hacky and wouldn't work with private dependencies. Users continue to consume either `<module>-cpp2-types` or, if they need custom serialization, `<module>-cpp2-serialization`.

Add `cpp2_serialization_srcs` parameter to `thrift_library` for injecting C++ sources that depend on serialization. It is mostly useful for instantiating serialization code for additional protocols.

Add a bunch of missing includes to make autodeps pull in the correct dependencies (`<module>-cpp2-types` or `<module>-cpp2-serialization`).

Reviewed By: iahs

Differential Revision: D78697921

fbshipit-source-id: 5da489eb7c5549bbab4331a6cd4557a6c19cf5bb