SimpleXMLElement::isTruthy
Summary:
Add a method that captures the behavior of coercing a SimpleXMLElement object to a bool.

PHP docs describe this behavior as "SimpleXML objects created from empty elements without attributes." (https://www.php.net/manual/en/language.types.boolean.php#language.types.boolean.casting)

The purpose is this function is to have a way to explicitly convert the object into a bool in a way specific to the type so that we can eventually ban arbitrary boolean coercions.

Reviewed By: paulbiss

Differential Revision: D80103573

fbshipit-source-id: 1a337badc642edc23947563ed78200cb420b59be