Type static method pointers as though there were an implicit class parameter
Summary: Updates the implementation of `extract_static_method` so simulate the prescence of an implicit `class<C<T>>` parameter. This is only necessary when we are removing redundant generics - before this diff, if a class-level generic appears only covariantly we would replace the generic with it's lower bound. After this diff, we make no subsititution since the parameter appears in the implicit parameter in a contravariant position

Reviewed By: andrewjkennedy

Differential Revision: D80167839

fbshipit-source-id: 57984a1fc29fb8d3bb129fcee32b507f5bb8a71f