Add ResourcePoolMigrationHelper.
Summary:
This gives us a single function to use during resource pool migration rollouts.

It provides:
1. A gflag (`--thrift_disable_resource_pool_migration`) that service owners can set to stop our migration in emergencies.
2. A thrift flag (`enable_resource_pool_migration`) that we will slowly enable for build rules (see parent diff D79852307).

Following diffs will check the `ResourcePoolMigrationHelper::isMigrationEnabled` (for `true`) to enable new, migrated behavior or (for `false`) to enable old, non-migrated behavior. Once the old behavior is completely disabled everywhere, it will be deleted and this helper will be removed.

If the migration is enabled and this migration helper is still compiled into code, it will log a warning telling service owners how to roll back the migration.

Reviewed By: tlj77

Differential Revision: D79852347

fbshipit-source-id: 7a7cc781d5f927068e907c075fa72e66db645fdd