fix sink exception type on client side
Summary:
When the client sink generator throws an exception, we expect two things to happen, per the cpp2 reference implementation:
1. exception propagated to server and raises in generator (in appropriate sequence)
2. exception raised to client

The issue with current implementation is that the original structured exception is recoverable from the wrapped `PythonUserException`, but that's not possible because the cpp2 `ClientSink::sink` method is wrapping it in an exception called `SinkThrew` that discards the IOBuf we need to recover the fields.

So propose adding a template param to `ClientSink::sink` so we can avoid wrapping the `PythonUserException`. That way, it gets raised and we can re-hyrdate it in the Python sink final response callback.

Reviewed By: prakashgayasen

Differential Revision: D79934887

fbshipit-source-id: 583abb19b7a3025da596f068b8a439aa479004ab