Enable multi-line mode in hphpd with 'k'
Summary:
`fbdbg` uses the command `k` to allow multi-line inputs. It would be nice to also have this available for `hphpd`. This diff enables also using `k` for multi-line inputs.

This diff also lets us abort multi-line mode with ctrl+c.

 https://pxl.cl/7Tc4d

I updated the expect files as well, so the core of this diff are the changes in `fbcode/hphp/runtime/debugger`.

Reviewed By: nt591

Differential Revision: D80119982

fbshipit-source-id: ce96f84403799c514c5219582ba0b85f62ec44f0