Delete CmdException and change 'c' to be 'constant' instead of 'continue'
Summary:
`fbdbg` uses the command `k` to allow multi-line inputs. It would be nice to also have this available for `hphpd`.

Right now, `k` in `hphpd` is used for `konstant`, since `c` is used for `continue`. Since we don't have breakpoints, the `continue` command is redundant, and we can stop using it. I didn't delete the code since it might be nice to keep for Hack notebooks.

 https://pxl.cl/7Tc4d
I updated the expect files as well, so the core of this diff are the changes in `fbcode/hphp/runtime/debugger`.

Reviewed By: nt591

Differential Revision: D80119983

fbshipit-source-id: 808997cd0c2fce8909fda99334cc96ec407c3512