allow serializing sb profile multiple times
Summary: I don't see a good reason why we had it in the first place. If we want to prevent it from happening, we can manage that in the userland. This function is particularly useful if someone wants to look into the sandbox tc, I wanna run it many times.

Differential Revision: D80664136

fbshipit-source-id: 821bd424ea3a64818841d5a80107e92f8d386336