Add option to disable 16kb buffering policy in Client streaming
Summary:
Thrift client sends more credits to server after receiving either `chunkBuffersize/2` payloads or 16kb data.

When the size of each payload is 16kb or close to it, client will send 1 credit after every payload. This increases the number of times we send credits -> thus increasing IO.

For MySQL client, after adding additional server and client side logs, it seems that Hack client is very slow and there seems to be constant IO pressure.
Since all MySQL streaming chunks are 16kb, the current theory is that constant IO is due to constantly sending 1 credit.

This diff adds an extension method to disable 16kb buffering.  Once this is released to www, we will add the necessary changes on the client side for MySQL client.

This policy was investigated during a previous issue as well. At the time, we had determined that removing this policy on c++ client has substantial difference but doesn't have the same impact on hack side.

But it's possible that the hack side impact can only be seen in production when there are multiple requests running using the same IO threads.

Reviewed By: tlj77

Differential Revision: D81055765

fbshipit-source-id: c53b029ef097f66527b8b3d0daf05ce0766acceb