Keep module interface docs in interfaces
Summary:
## Before this change

We followed **4 different conventions** for where doc comments for module interfaces go:
- (1) interface file file
- (2) Both interface file and implementation file, exact same comments in each
    - I suspect we duplicated doc comments due to a prior limitation of oxidization **that's fixed recently** (D79951959).
- (3) Both interface file and implementation file, slightly different comments in each
- (4) Implementation file only

## After this change

**1 dominant convention**: doc comments for module interfaces in the interface files

> for as many files as I could get to before I got tired.

## Why document interfaces in interface files

**Practical**:
- When we do go to definition, we go to the .mli. So comments in the .mli are more discoverable.
- Having to update comments in two places is verbose, mistake-prone, and annoying.

**Rule-based**:
- Our internal style guide says to follow Jane Street style and this is how Jane Street does it.

**Consistency**:
- One way is better than four.
- This is the OCaml ecosystem's way to do it, and I like doing things the normal way for a programming language ecosystem unless there's a strong reason not to.

## Useful Keybindings

> I include these even though they're probably well-known, just in case it's helpful

I use cmd-shift-a to switch between .mli and .ml. fwiw I find it really helpful to keep track of what the public interface of a module is and what the overall shape of the API is. I frequently look at .mlis and really appreciate it when that's where the doc comments are.
 {F1981072104}

I also use cmd-shift-e to reveal the active file in the explorer. That way I can see what related modules might be. I think this behhavior enabled by default, but just in case it's not:

 {F1981072110}

Reviewed By: geralt-encore

Differential Revision: D79949390

fbshipit-source-id: 3326bdbe1fe2e079fc2c8baf2b7c2ea4488cfdca