Fix positions of type applications
Summary:
The generic `get_pos` helper over types (calls Reason.to_pos) is used mainly for reporting user errors. It reports a suboptimal position in the case. where we have a type application, usually showing a definition rather than the application.

More seriously, this affects the check applied to inherited trait properties as the check depends on the position.
https://www.internalfb.com/code/fbsource/[5e992a84ae1af93a1c27143135a2fdba00e120b8]/fbcode/hphp/hack/src/typing/typing_class.ml?lines=951-961
#fileatask rework generic static property check to be on original def, not on substitution (for traits)

Reviewed By: mheiber

Differential Revision: D80110617

fbshipit-source-id: 3d524fc881179543f8ebf63aa276056f97528612