skip indexing files over 5MB
Summary:
Context: for now we maintain manually a list of files to block list in the hack indexer. The indexer periodically breaks and we need to add new files. Upon investigation all the blocked files but one are above 5MB.

This diff programatically skips files above 5MB.

The 5MB limit is hard-coded for now, it could be made a config.

The following scuba queries show that files above 5MB are about 42/6.5M, less than 0.00001%.
* https://fburl.com/scuba/hh_process_file_history/jk98zp4g
* https://fburl.com/scuba/hh_process_file_history/qsbibz9b

We've also tried setting a time out per file in D80700417, but timing out breaks some invariant in the decl heap which causes crashes. This is a safer solution which is also faster.

Reviewed By: jjuliamolin, phlalx

Differential Revision: D81242579

fbshipit-source-id: 8b1eaf5c10a72874603e374cd888da9c0c328120