Reproduce type-check failure with enum conversion to thrift-py-deprecated
Summary:
# What?
Reproduce the issue that the use of `thrift_enums._to_py_deprecated()` causes

```
expected `some.py_deprecated.enum` but got `int'
```

# What does the test reveal?
The test shows that the run-time behavior is correct and the issue is a static type analysis failure caused by an incorrect type annotation in generated code.

# What's next?
A subsequent change resolves the type-check error, and removes the `pyre-fixme`.

Reviewed By: yfeldblum, prakashgayasen

Differential Revision: D81289541

fbshipit-source-id: 7786d298477a4662723bc45b48bf9d461c481159