propagate position information in is_check for better debugging of T235374244
Summary:
## What

propagate position information in is_check

## Why

For better debugging of T235374244

Previous attempt {D80710829} did not improve logging. I suspect we have some laziness via callbacks and so we're escaping the `with` introduced in D80710829.

## Considerations

I'm not sure why we use `Pos.none` a lot in `is_check.ml`
- Given our type inference is non-local, maybe the author was worried about misleading signal
- However, I suspect it's better to have *some* position information rather than none. Maybe the lack of position information is why our telemetry for this error is missing position information (see fb-only section)

Reviewed By: geralt-encore

Differential Revision: D81488483

fbshipit-source-id: 161bccda7d8d2c4307808bcf03c8f8becba5cdbf